+++ app/src/main/java/com/techcourse/controller/LogoutController.java
@@ -11,9 +11,8 @@ import webmvc.org.springframework.web.servlet.view.JspView;
 @Controller
 public class LogoutController {
 
-    @RequestMapping(value = "/logout", method = RequestMethod.POST)
-    public ModelAndView logout(final HttpServletRequest request,
-                               final HttpServletResponse response) {
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public ModelAndView logout(HttpServletRequest request, HttpServletResponse response) {
         final var session = request.getSession();
         session.removeAttribute(UserSession.SESSION_KEY);
         return new ModelAndView(new JspView("redirect:/"));
