+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/HandlerAdapters.java
@@ -1,7 +1,4 @@
-package com.techcourse;
-
-import jakarta.annotation.Nonnull;
-import webmvc.org.springframework.web.servlet.mvc.HandlerAdapter;
+package webmvc.org.springframework.web.servlet.mvc;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,7 +11,6 @@ public class HandlerAdapters {
         handlerAdapters.add(handlerAdapter);
     }
 
-    @Nonnull
     public HandlerAdapter getHandlerAdapter(Object handler) {
         return handlerAdapters.stream()
                 .filter(adapter -> adapter.supports(handler))
